pyproject.toml
[project]
name = "stock-prediction"
version = "0.1.1"
description = "Stock price prediction system using machine learning"
authors = [
    { name = "lin", email = "hvtga447@gmail.comw" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.0",
    "pandas==2.1.4",
    "scikit-learn==1.3.2",
    "matplotlib==3.8.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.4",
    "aiosignal==1.3.2",
    "akracer==0.0.13",
    "akshare==1.16.95",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "bs4==0.0.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "et-xmlfile==2.0.0",
    "fonttools==4.58.1",
    "frozenlist==1.6.0",
    "html5lib==1.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "joblib==1.5.1",
    "jsonpath==0.82.2",
    "kiwisolver==1.4.8",
    "lxml==5.4.0",
    "multidict==6.4.4",
    "nest-asyncio==1.6.0",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pillow==11.2.1",
    "pluggy==1.6.0",
    "propcache==0.3.1",
    "py-mini-racer==0.6.0",
    "pyparsing==3.2.3",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "scipy==1.15.3",
    "simplejson==3.20.1",
    "six==1.17.0",
    "soupsieve==2.7",
    "tabulate==0.9.0",
    "threadpoolctl==3.6.0",
    "tqdm==4.67.1",
    "tushare==1.4.21",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "xlrd==2.0.1",
    "yarl==1.20.0",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"
